refactor: unify MockUserLookup into shared test helper package
Consolidate duplicated MockUserLookup structs from commands/add_test.go
and policy/policyloader_test.go into a single struct in test/testutil/.

The unified MockUserLookup supports both single-user and multi-user
lookup scenarios via User (default) and Users (map) fields.

Local type aliases keep existing test code concise.

Closes #439
