+++ commands/add_test.go
@@ -17,33 +17,19 @@
 package commands
 
 import (
-	"os/user"
 	"testing"
 
 	"github.com/openpubkey/opkssh/policy"
 	"github.com/openpubkey/opkssh/policy/files"
+	"github.com/openpubkey/opkssh/test/testutil"
 	"github.com/spf13/afero"
 	"github.com/stretchr/testify/require"
 )
 
-// Duplicates code from multipolicyloader_test.go
-type MockUserLookup struct {
-	// User is returned on any call to Lookup() if Error is nil
-	User *user.User
-	// Error is returned on any call to Lookup() if non-nil
-	Error error
-}
+// MockUserLookup is an alias for testutil.MockUserLookup to keep test code concise.
+type MockUserLookup = testutil.MockUserLookup
 
-// Lookup implements policy.UserLookup
-func (m *MockUserLookup) Lookup(username string) (*user.User, error) {
-	if m.Error == nil {
-		return m.User, nil
-	} else {
-		return nil, m.Error
-	}
-}
-
-var ValidUser *user.User = &user.User{HomeDir: "/home/foo", Username: "foo"}
+var ValidUser = testutil.ValidUser
 
 func MockAddCmd(mockFs afero.Fs) *AddCmd {
 	mockUserLookup := &MockUserLookup{User: ValidUser}
